lmb-1331: move fetching history to the mandataris route
