+++ app/components/mandatarissen/history.hbs
@@ -1,41 +1,37 @@
-<div {{did-update this.onMandatarisUpdate @mandataris}}>
-  {{#if this.fetchHistory.isRunning}}
-    <Skeleton::Table @columns={{7}} @rows={{@mandatarissen.length}} />
-  {{else}}
-    <AuTable>
-      <:header>
+<div>
+  <AuTable>
+    <:header>
+      <tr>
+        <th>Van</th>
+        <th>Tot</th>
+        <th>Status</th>
+        {{#if this.toonBeleidsdomeinen}}
+          <th>#Beleidsdomeinen</th>
+        {{/if}}
+        {{#if this.toonRangorde}}
+          <th>Rangorde</th>
+        {{/if}}
+        <th>Fractie</th>
+        <th>Aangepast door</th>
+        <th>Op</th>
+        <th></th>
+      </tr>
+    </:header>
+    <:body>
+      {{#each @history as |historyItem|}}
+        <Mandatarissen::HistoryRow
+          @mandataris={{historyItem.mandataris}}
+          @selected={{historyItem.selected}}
+          @corrections={{historyItem.corrections}}
+          @toonBeleidsdomeinen={{this.toonBeleidsdomeinen}}
+          @toonRangorde={{this.toonRangorde}}
+        />
+      {{else}}
         <tr>
-          <th>Van</th>
-          <th>Tot</th>
-          <th>Status</th>
-          {{#if this.toonBeleidsdomeinen}}
-            <th>#Beleidsdomeinen</th>
-          {{/if}}
-          {{#if this.toonRangorde}}
-            <th>Rangorde</th>
-          {{/if}}
-          <th>Fractie</th>
-          <th>Aangepast door</th>
-          <th>Op</th>
-          <th></th>
+          <td colspan="4">Geen aanpassingen gevonden</td>
         </tr>
-      </:header>
-      <:body>
-        {{#each this.history as |historyItem|}}
-          <Mandatarissen::HistoryRow
-            @mandataris={{historyItem.mandataris}}
-            @selected={{historyItem.selected}}
-            @corrections={{historyItem.corrections}}
-            @toonBeleidsdomeinen={{this.toonBeleidsdomeinen}}
-            @toonRangorde={{this.toonRangorde}}
-          />
-        {{else}}
-          <tr>
-            <td colspan="4">Geen aanpassingen gevonden</td>
-          </tr>
-        {{/each}}
-      </:body>
-    </AuTable>
-    {{yield}}
-  {{/if}}
+      {{/each}}
+    </:body>
+  </AuTable>
+  {{yield}}
 </div>
