+++ app/templates/mandatarissen/persoon/mandataris.hbs
@@ -117,6 +117,7 @@
         @mandataris={{@model.mandataris}}
         @mandatarissen={{@model.mandatarissen}}
         @form={{@model.mandatarisEditForm}}
+        @history={{@model.history}}
       />
     </ul>
   </div>
